requirements.txt
ultralytics
onnx
onnxslim
onnxruntime
pyyaml